refine random state in exported codes #933
